+++ doc/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support for Wild Worlds
 - Support for Mutants and Masterminds
 - Support for Mothership
+- Support for additional Exalted dice rolls
 
 ## [1.4.0] - 2025-07-05